requirements.txt
PyQt5
PyGithub
google-generativeai